+++ apps/ledger-live-mobile/package.json
@@ -183,7 +183,6 @@
     "@ledgerhq/live-dmk-shared": "workspace:^",
     "@ledgerhq/live-dmk-speculos": "workspace:^",
     "@ledgerhq/live-engagement": "workspace:^",
-    "@ledgerhq/live-hooks": "workspace:^",
     "@ledgerhq/live-network": "workspace:^",
     "@ledgerhq/live-signer-evm": "workspace:^",
     "@ledgerhq/live-wallet": "workspace:^",
