+++ libs/ledger-live-common/package.json
@@ -349,7 +349,6 @@
     "@ledgerhq/live-currency-format": "workspace:^",
     "@ledgerhq/live-dmk-shared": "workspace:^",
     "@ledgerhq/live-env": "workspace:^",
-    "@ledgerhq/live-hooks": "workspace:*",
     "@ledgerhq/live-network": "workspace:^",
     "@ledgerhq/live-promise": "workspace:^",
     "@ledgerhq/live-signer-aleo": "workspace:^",
