+++ libs/live-countervalues-react/package.json
@@ -24,7 +24,6 @@
     "@ledgerhq/types-live": "workspace:*",
     "@ledgerhq/live-countervalues": "workspace:*",
     "@ledgerhq/ledger-wallet-framework": "workspace:^",
-    "@ledgerhq/live-hooks": "workspace:*",
     "bignumber.js": "9"
   },
   "devDependencies": {
