--- libs/live-hooks/package.json
@@ -1,86 +1,0 @@
-{
-  "name": "@ledgerhq/live-hooks",
-  "version": "0.6.0",
-  "description": "Ledger Live React hooks",
-  "keywords": [
-    "Ledger"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/LedgerHQ/ledger-live.git"
-  },
-  "bugs": {
-    "url": "https://github.com/LedgerHQ/ledger-live/issues"
-  },
-  "homepage": "https://github.com/LedgerHQ/ledger-live/tree/develop/libs/env",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "lib/index.js",
-  "module": "lib-es/index.js",
-  "types": "lib/index.d.ts",
-  "license": "Apache-2.0",
-  "peerDependencies": {
-    "react": ">=16"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@ledgerhq/test-quarantine": "workspace:*",
-    "@types/node": "catalog:",
-    "@types/jest": "catalog:",
-    "@testing-library/dom": "catalog:",
-    "@testing-library/react": "catalog:",
-    "jest": "catalog:",
-    "react": "catalog:",
-    "react-dom": "catalog:",
-    "jest-environment-jsdom": "catalog:",
-    "@swc/jest": "catalog:",
-    "@swc/core": "catalog:"
-  },
-  "scripts": {
-    "clean": "rimraf lib lib-es",
-    "build": "tsc --project tsconfig.build.json && tsc --project tsconfig.build.json -m esnext --moduleResolution bundler --outDir lib-es",
-    "prewatch": "pnpm build",
-    "watch": "tsc --watch --project tsconfig.build.json",
-    "watch:es": "tsc --watch --project tsconfig.build.json -m esnext --moduleResolution bundler --outDir lib-es",
-    "lint": "oxlint -c ../oxc-live-libs/.oxlintrc.json ./src",
-    "lint:fix": "pnpm lint --fix",
-    "typecheck": "tsc --noEmit --customConditions node",
-    "unimported": "pnpm knip --directory ../.. -W libs/live-hooks",
-    "test": "jest",
-    "coverage": "jest --coverage"
-  },
-  "typesVersions": {
-    "*": {
-      "*.json": [
-        "*.json"
-      ],
-      "*": [
-        "lib/*"
-      ],
-      "lib/*": [
-        "lib/*"
-      ],
-      "lib-es/*": [
-        "lib-es/*"
-      ]
-    }
-  },
-  "exports": {
-    "./lib/*": "./lib/*.js",
-    "./lib/*.js": "./lib/*.js",
-    "./lib-es/*": "./lib-es/*.js",
-    "./lib-es/*.js": "./lib-es/*.js",
-    "./*": {
-      "@ledgerhq/source": "./src/*.ts",
-      "require": "./lib/*.js",
-      "default": "./lib-es/*.js"
-    },
-    "./*.js": {
-      "@ledgerhq/source": "./src/*.ts",
-      "require": "./lib/*.js",
-      "default": "./lib-es/*.js"
-    },
-    "./package.json": "./package.json"
-  }
-}
