--- libs/live-hooks/tsconfig.build.json
@@ -1,20 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "customConditions": [],
-    "types": []
-  },
-  "exclude": [
-    "**/*.test.ts",
-    "**/*.test.tsx",
-    "**/*.spec.ts",
-    "**/*.spec.tsx",
-    "**/__tests__/**/*",
-    "**/tests/**/*",
-    "**/__mocks__/**/*",
-    "**/*.test.js",
-    "**/*.test.jsx",
-    "**/*.spec.js",
-    "**/*.spec.jsx"
-  ]
-}
