--- libs/live-hooks/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "../../tsconfig.base",
-  "compilerOptions": {
-    "declaration": true,
-    "declarationMap": true,
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "jsx": "react",
-    "lib": ["es2020", "dom"],
-    "outDir": "lib",
-    "rootDir": "src",
-    "paths": {
-      "*": ["./*"],
-      "react": ["../../node_modules/@types/react"],
-      "@types/react": ["../../node_modules/@types/react"]
-    },
-    "types": ["node", "jest"]
-  },
-  "include": ["src/**/*"]
-}
